Develop tables component, set up for formatting changes
