+++ src/components/Header.js
@@ -1,6 +1,6 @@
 import React from "react"
 
-function Header() {
+function Header({dataTitle}) {
 
 
     return (
